requirements.txt
argon2==0.1.10
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.15.1
Django==4.2.3
djangorestframework==3.14.0
Pillow==10.0.0
psycopg==3.1.9
pycparser==2.21
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
